requirements.txt
app>=0.0.1
config>=0.5.1
Flask>=2.2.5
Flask_Login>=0.6.2
Flask_Mail>=0.9.1
flask_migrate>=4.0.4
flask_sqlalchemy>=3.0.5
Flask_WTF>=1.1.1
geonamescache>=2.0.0
locationtagger>=0.0.1
lxml_html_clean>=0.1.1
newsapi_python>=0.2.7
nltk>=3.8.1
pandas>=1.5.3
PyJWT>=2.8.0
Requests>=2.31.0
spacy>=3.5.0
vaderSentiment>=3.3.2
Werkzeug>=2.3.8
WTForms>=3.0.1
WTForms_Components>=0.10.5
